requirements.txt
requests
paho-mqtt
numpy
os